Fix child signal handling
When a child recieves a signal, the parent has the option to hide the signal.
This passes the signal number to the child process using `ptrace` so that the
child will handle it (if it can). This should avoid an infinite loop in the
caller when the child hits a segmentation fault.